Промежуточный коммит - тесты

+++ src/main/java/org/mnuykin/dto/rq/CommentCreateRqDto.java
@@ -6,5 +6,5 @@ public record CommentCreateRqDto (
         @NotNull
         String text,
         @NotNull
-        Integer postId) {
+        Long postId) {
 }
